+++ src/colony/base/__init__.py
@@ -36,7 +36,16 @@ from . import test
 from . import util
 from . import vars
 
-from .config import conf, conf_prefix, conf_suffix, conf_s, conf_r, conf_d, conf_ctx
+from .config import (
+    conf,
+    conf_prefix,
+    conf_suffix,
+    conf_s,
+    conf_r,
+    conf_d,
+    conf_ctx,
+    conf_override,
+)
 from .decorators import (
     load_plugin,
     plugin_meta_information,